Moved test code
added homepage and chimedemo page and moved test code out of app
