+++ package.json
@@ -15,7 +15,8 @@
     "aws-amplify": "^6.6.6",
     "axios": "^1.7.8",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^7.0.1"
   },
   "devDependencies": {
     "@aws-amplify/backend": "^1.5.0",
