+++ src/pages/chimedemo.tsx
@@ -1,0 +1,10 @@
+
+const Chimedemo = () => {
+    return (
+        <div>
+            <h1>Welcome to the Chimedemo Page</h1>
+        </div>
+    );
+};
+
+export default Chimedemo;
